+++ runtime/src/accounts_hash.rs
@@ -1580,7 +1580,7 @@ pub mod tests {
                     let packaged_result: ExpectedType = (
                         human_readable,
                         is_last_slice,
-                        lamports2 as u64,
+                        lamports2,
                         hash_result_as_string,
                     );
                     assert_eq!(expected[expected_index], packaged_result);
@@ -1651,7 +1651,7 @@ pub mod tests {
 
         let vecs = vec![vec![account_maps.to_vec()]];
         let result = test_de_dup_accounts_in_parallel(&vecs);
-        assert_eq!(result, (vec![&val.hash], val.lamports as u64, 1));
+        assert_eq!(result, (vec![&val.hash], val.lamports, 1));
 
         // zero original lamports, higher version
         let val = CalculateHashIntermediate::new(hash, ZERO_RAW_LAMPORTS_SENTINEL, key);
